+++ src/Utility/TextSplitter.cpp
@@ -7,7 +7,7 @@ int TextSplitter::rangeWidth(long long int textLength) {
 // It's hard cut the text into substrings of equal size without getting out of range sometimes,
 // so when we get out of range a <-1, -1> is pushed back to indicate that this tuple is wrong and thread should not process it
 
-std::vector<std::tuple<long long int, long long int>>& TextSplitter::splitText(long long int textLength, int patternLength) {
+std::vector<std::tuple<size_t, size_t>>& TextSplitter::splitText(long long int textLength, int patternLength) {
     int width = rangeWidth(textLength);
 
     for(int i = 0; i < textLength; i += width) {
